+++ eslint.config.js
@@ -2,20 +2,29 @@ import js from '@eslint/js'
 import globals from 'globals'
 import reactHooks from 'eslint-plugin-react-hooks'
 import reactRefresh from 'eslint-plugin-react-refresh'
+import tseslint from '@typescript-eslint/eslint-plugin'
+import tsparser from '@typescript-eslint/parser'
 import { defineConfig, globalIgnores } from 'eslint/config'
 
 export default defineConfig([
   globalIgnores(['dist']),
   {
-    files: ['**/*.{js,jsx}'],
+    files: ['**/*.{ts,tsx}'],
     extends: [
       js.configs.recommended,
       reactHooks.configs.flat.recommended,
       reactRefresh.configs.vite,
     ],
+    plugins: {
+      '@typescript-eslint': tseslint,
+    },
     languageOptions: {
       globals: globals.browser,
+      parser: tsparser,
       parserOptions: { ecmaFeatures: { jsx: true } },
+    },
+    rules: {
+      ...tseslint.configs.recommended.rules,
     },
   },
 ])